+++ routes/web.php
@@ -1,6 +1,7 @@
 <?php
 
 use Illuminate\Support\Facades\Route;
+use App\Http\Controllers\AuthController;
 use App\Http\Controllers\ClassController;
 use App\Http\Controllers\StudentController;
 use App\Http\Controllers\TeacherController;
@@ -18,34 +19,34 @@ use App\Http\Controllers\ExtracurricularController;
 */
 
 Route::get('/', function () {
-    return view('home',[
-    'name' => 'Haikal', 
-    'role' => 'admin',
-    'buah' => ['apel', 'jeruk', 'mangga'],
-    ]);
-});
+    return view('home');
+})->middleware('auth');
 
-Route::get('/students', [StudentController::class, 'index']);
-Route::get('/student/{id}', [StudentController::class, 'show']);
-Route::get('/student-tambah', [StudentController::class, 'create']);
-Route::post('/student', [StudentController::class, 'store']);
-Route::get('/student-edit/{id}', [StudentController::class, 'edit']);
-Route::put('/student/{id}', [StudentController::class, 'update']);
-Route::get('/student-delete/{id}', [StudentController::class, 'delete']);
-Route::delete('/student-hapus/{id}', [StudentController::class, 'destroy']);
-Route::get('/student-terhapus', [StudentController::class, 'deletedStudent']);
-Route::get('/student/{id}/restore', [StudentController::class, 'restore']);
+Route::get('/login', [AuthController::class, 'login'])->name('login')->middleware('guest');
+Route::post('/login', [AuthController::class, 'authenticating'])->middleware('guest');
+Route::get('/logout', [AuthController::class, 'logout'])->middleware('auth');
+
+Route::get('/students', [StudentController::class, 'index'])->middleware('auth');
+Route::get('/student/{id}', [StudentController::class, 'show'])->middleware(['auth','only-admin-teacher']);
+Route::get('/student-tambah', [StudentController::class, 'create'])->middleware(['auth','only-admin-teacher']);
+Route::post('/student', [StudentController::class, 'store'])->middleware(['auth','only-admin-teacher']);
+Route::get('/student-edit/{id}', [StudentController::class, 'edit'])->middleware(['auth','only-admin-teacher']);
+Route::put('/student/{id}', [StudentController::class, 'update'])->middleware(['auth','only-admin-teacher']);
+Route::get('/student-delete/{id}', [StudentController::class, 'delete'])->middleware(['auth','only-admin']);
+Route::delete('/student-hapus/{id}', [StudentController::class, 'destroy'])->middleware(['auth','only-admin']);
+Route::get('/student-terhapus', [StudentController::class, 'deletedStudent'])->middleware(['auth','only-admin']);
+Route::get('/student/{id}/restore', [StudentController::class, 'restore'])->middleware(['auth','only-admin']);
 
 
 
-Route::get('/class', [ClassController::class, 'index']);
-Route::get('/class-detail/{id}', [ClassController::class, 'show']);
+Route::get('/class', [ClassController::class, 'index'])->middleware('auth');
+Route::get('/class-detail/{id}', [ClassController::class, 'show'])->middleware('auth');
 
-Route::get('/extracurricular', [ExtracurricularController::class, 'index']);
-Route::get('/eskul-detail/{id}', [ExtracurricularController::class, 'show']);
+Route::get('/extracurricular', [ExtracurricularController::class, 'index'])->middleware('auth')->middleware('auth');
+Route::get('/eskul-detail/{id}', [ExtracurricularController::class, 'show'])->middleware('auth');
 
-Route::get('/teacher', [TeacherController::class, 'index']);
-Route::get('/teacher-detail/{id}', [TeacherController::class, 'show']);
+Route::get('/teacher', [TeacherController::class, 'index'])->middleware('auth')->middleware('auth');
+Route::get('/teacher-detail/{id}', [TeacherController::class, 'show'])->middleware('auth');
